requirements.txt
Django==2.2.3
djangorestframework==3.11.0
pytz==2019.1
sqlparse==0.3.0
